allow whitelisting of specific AzureAD / local domain groups

+++ leanLAPS/leanLAPS.ps1
@@ -21,6 +21,11 @@ $removeOtherLocalAdmins = $False #if set to True, will remove ALL other local ad
 $onlyRunOnWindows10 = $True #buildin protection in case an admin accidentally assigns this script to e.g. a domain controller
 $markerFile = Join-Path $Env:TEMP -ChildPath "leanLAPS.marker"
 $markerFileExists = (Test-Path $markerFile)
+$approvedAdmins = @( #specify names for Local AD groups or SID's for Azure groups such as Global Admins and Device Administrators. These are specific to your tenant, you can get them on a device by running: ([ADSI]::new("WinNT://$($env:COMPUTERNAME)/$((New-Object System.Security.Principal.SecurityIdentifier("S-1-5-32-544")).Translate([System.Security.Principal.NTAccount]).Value.Split("\")[1]),Group")).Invoke('Members') | % {([ADSI]$_).Path.Split("/")[-1]}
+"S-1-12-1-3524033542-1298357157-3149679747-1970349937"
+"S-1-12-1-3862487121-1222458618-970102617-4043066380"
+"Domain Admins"
+)
 
 
 function Get-RandomCharacters($length, $characters) { 
@@ -111,10 +116,12 @@ try{
     #remove other local admins if specified, only executes if adding the new local admin succeeded
     if($removeOtherLocalAdmins){
         foreach($administrator in $administrators){
-            if($administrator -ne $localAdminName){
+            if($administrator -ne $localAdminName -and $approvedAdmins -notcontains $administrator){
                 Write-CustomEventLog "removeOtherLocalAdmins set to True, removing $($administrator) from Local Administrators"
                 $res = Remove-LocalGroupMember -Group $administratorsGroupName -Member $administrator -Confirm:$False
                 Write-CustomEventLog "Removed $administrator from Local Administrators"
+            }else{
+                Write-CustomEventLog "Not removing $($administrator) because of whitelisting"
             }
         }
     }else{
